refactor: 소셜타입 방어 로직 분리
