+++ src/domains/password-edit/container/PasswordEditContainer.tsx
@@ -1,18 +1,10 @@
-import React, { useEffect } from "react";
-import { useSelector } from "react-redux";
-import { useNavigate } from "react-router-dom";
-import Path from "routes/Path";
-import { userSelector } from "stores/user";
 import { PasswordEditTemplate } from "../components";
 import PasswordEditForm from "../components/PasswordEditForm";
 import PasswordEditHead from "../components/PasswordEditHead";
+import useCheckSocialType from "../hooks/useCheckSocialType";
 
 function PasswordEditContainer() {
-  const { socialType } = useSelector(userSelector);
-  const navigate = useNavigate();
-  useEffect(() => {
-    if (socialType === "google") navigate(Path.MyPage);
-  }, [navigate, socialType]);
+  useCheckSocialType();
 
   return (
     <PasswordEditTemplate>
